+++ graylog2-server/src/main/java/org/graylog/plugins/pipelineprocessor/PipelineProcessorModule.java
@@ -21,6 +21,7 @@ import org.graylog.plugins.pipelineprocessor.audit.PipelineProcessorAuditEventTy
 import org.graylog.plugins.pipelineprocessor.functions.ProcessorFunctionsModule;
 import org.graylog.plugins.pipelineprocessor.periodical.LegacyDefaultStreamMigration;
 import org.graylog.plugins.pipelineprocessor.processors.PipelineInterpreter;
+import org.graylog.plugins.pipelineprocessor.processors.PipelineInterpreterState;
 import org.graylog.plugins.pipelineprocessor.rest.PipelineConnectionsResource;
 import org.graylog.plugins.pipelineprocessor.rest.PipelineResource;
 import org.graylog.plugins.pipelineprocessor.rest.PipelineRestPermissions;
@@ -47,7 +48,7 @@ public class PipelineProcessorModule extends PluginModule {
                 PipelineProcessorMessageDecorator.class,
                 PipelineProcessorMessageDecorator.Factory.class);
 
-        install(new FactoryModuleBuilder().build(PipelineInterpreter.State.Factory.class));
+        install(new FactoryModuleBuilder().build(PipelineInterpreterState.Factory.class));
 
         addAuditEventTypes(PipelineProcessorAuditEventTypes.class);
     }
